Week 1 debug
1. Player & Monster's attack combine

+++ Assets/Scripts/Creature.cs
@@ -15,7 +15,7 @@ public class Creature : MonoBehaviour {
 
     public float MaxHP { get; set; }
 
-    public void Move(int feet)
+    public void Move(float feet)
     {
         BoardController.distance += feet;
         Debug.Log("Distance is " + BoardController.distance);
